+++ db/migrate/20200330145039_remove_finance_from_activities.rb
@@ -1,0 +1,5 @@
+class RemoveFinanceFromActivities < ActiveRecord::Migration[6.0]
+  def change
+    remove_column :activities, :finance, :string
+  end
+end