+++ spec/features/staff/users_can_edit_an_activity_spec.rb
@@ -246,14 +246,6 @@ def assert_all_edit_links_go_to_the_correct_form_step(activity:)
   end
   click_on(I18n.t("generic.link.back"))
 
-  within(".finance") do
-    click_on(I18n.t("generic.link.edit"))
-    expect(page).to have_current_path(
-      activity_step_path(activity, :finance)
-    )
-  end
-  click_on(I18n.t("generic.link.back"))
-
   within(".aid_type") do
     click_on(I18n.t("generic.link.edit"))
     expect(page).to have_current_path(